Migrate enzyme Url tests to React Testing Library
